add admin

+++ api/database/migrations/2023_05_13_230028_create_paiements_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('paiements', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('id_intervenant');
-            $table->integer('id_etab');
+            $table->unsignedBigInteger('id_etab');
             $table->integer('VH');
             $table->float('taux_h');
             $table->float('brut')->nullable();
